24-25_Code/SOAR_Echo_Base/requirements.txt
Flask==3.0.0
Flask-SocketIO==5.3.6
pyserial==3.5